backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.2.1
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
httpx==0.27.0
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
isort==5.13.2
flake8==7.0.0
PyGithub==2.1.1
python-dotenv==1.0.0
